Modified JobState to make it more robust.
The script will now check every second for the job status instead
of checking it after 10 seconds. The script will also check the status
immediately after making the appropriate action (e.g. putting back the paper in tray)
without needing the tester to press enter. This will fix issues/scenarios from the
old implementation like:

1. In the old implementation, after 10 seconds, The script is expecting that
the job status should be 'in progress' after putting back the paper in the tray.
This will work if printing is slow, but there are printers that could print
very fast so there are times that the job status has already transitioned to 'printed'
in less than 10 seconds. So the script will sometimes fail.

2. Tester wasn't able to press the enter button on time after putting back the paper on tray.
During this time, the job status has already been transitioned to
'Printed', so the script will now fail since it's expecting the status
to be 'in-progress' state.

3. If the internet connection is very slow, Job status will only be able
to transition into 'Printed' state after 30 - 60 seconds. So the script will also
fail because it is expecting that the state is already in 'Printed' state in just 10
seconds.

Also modified the Registration suite. Tester now needs to press enter
after registration is accepted. This is to avoid failures when the tester
failed to accept the registration in 20 seconds. Also increased the
timer to 1 minute.